ZeroClassic v2.0.20 "Wardenclyffe"

+++ src/snark/libsnark/zk_proof_systems/ppzksnark/r1cs_ppzksnark/r1cs_ppzksnark.tcc
@@ -312,7 +312,7 @@ r1cs_ppzksnark_keypair<ppT> r1cs_ppzksnark_generator(
 
     const Fr<ppT>      rC = rA * rB;
 
-    // consrtuct the same-coefficient-check query (must happen before zeroing out the prefix of At)
+    // construct the same-coefficient-check query (must happen before zeroing out the prefix of At)
     Fr_vector<ppT> Kt;
     Kt.reserve(qap_inst.num_variables()+4);
     for (size_t i = 0; i < qap_inst.num_variables()+1; ++i)
@@ -443,9 +443,9 @@ knowledge_commitment<T1, T2> r1cs_compute_proof_kc(const qap_witness<Fr<ppT> > &
 #endif
 
 #ifdef MULTICORE
-    const uint64_t chunks = omp_get_max_threads(); // to override, set OMP_NUM_THREADS env var or call omp_set_num_threads()
+    const size_t chunks = omp_get_max_threads(); // to override, set OMP_NUM_THREADS env var or call omp_set_num_threads()
 #else
-    const uint64_t chunks = 1;
+    const size_t chunks = 1;
 #endif
 
     returnval = returnval + kc_multi_exp_with_mixed_addition<T1, T2, Fr<ppT> >(
@@ -471,9 +471,9 @@ G1<ppT> r1cs_compute_proof_K(const qap_witness<Fr<ppT>> &qap_wit, const G1_vecto
 #endif
 
 #ifdef MULTICORE
-    const uint64_t chunks = omp_get_max_threads(); // to override, set OMP_NUM_THREADS env var or call omp_set_num_threads()
+    const size_t chunks = omp_get_max_threads(); // to override, set OMP_NUM_THREADS env var or call omp_set_num_threads()
 #else
-    const uint64_t chunks = 1;
+    const size_t chunks = 1;
 #endif
 
     G1<ppT> g_K = K_query[0] + zk_shift;
@@ -500,9 +500,9 @@ G1<ppT> r1cs_compute_proof_H(const qap_witness<Fr<ppT> > &qap_wit, const G1_vect
 #endif
 
 #ifdef MULTICORE
-    const uint64_t chunks = omp_get_max_threads(); // to override, set OMP_NUM_THREADS env var or call omp_set_num_threads()
+    const size_t chunks = omp_get_max_threads(); // to override, set OMP_NUM_THREADS env var or call omp_set_num_threads()
 #else
-    const uint64_t chunks = 1;
+    const size_t chunks = 1;
 #endif
 
     g_H = g_H + multi_exp<G1<ppT>, Fr<ppT> >(
@@ -685,7 +685,7 @@ bool r1cs_ppzksnark_online_verifier_weak_IC(const r1cs_ppzksnark_processed_verif
 
     const accumulation_vector<G1<ppT> > accumulated_IC = pvk.encoded_IC_query.template accumulate_chunk<Fr<ppT> >(primary_input.begin(), primary_input.end(), 0);
     const G1<ppT> &acc = accumulated_IC.first;
-    
+
     if (!proof.is_well_formed())
     {
         return false;
